requirements/biobrick-base.txt
django-haystack==2.6.1
elasticsearch==5.4.0
